# Extracts byte array specific code to ByteArrayBuffer