--- a/zipkin-server/src/main/java/zipkin2/server/internal/ZipkinQueryApiV2.java
+++ b/zipkin-server/src/main/java/zipkin2/server/internal/ZipkinQueryApiV2.java
@@ -136,7 +136,9 @@ public class ZipkinQueryApiV2 {
     }
 
     @Override public void write(String value, Buffer buffer) {
-      buffer.writeByte('"').writeUtf8(value).writeByte('"');
+      buffer.writeByte('"');
+      buffer.writeUtf8(value);
+      buffer.writeByte('"');
     }
   };