This starts transitioning the internals of our codec library so that
other buffers can be used directly.

The first step is to separate out certain things that have to be done
from how to do it with byte arrays. A few frequently used functions
are directly implemented with byte array and hopefully can be trivially
reimplemented with other types of buffers.

See #2435