@adriancole 2019-03-13
@anuraaga this should make future work a bit easier. Since these are internal classes we can reshape them a bit with each pull request, taking more care with the non-internal packages.