@adriancole 2019-03-13
running CodecBenchmarks before and after to ensure I didn't mess anything up