@adriancole 2019-03-14
This doesn't improve performance, and the runs are not reliable enough to say how much it slows things down. It seems to a little, but if we make a bytebuffer version and that saves allocating arrays, it would likely be worth it.

before
```
Benchmark                                        Mode  Cnt   Score   Error  Units
CodecBenchmarks.readChineseSpan_json             avgt   15   4.738 ± 0.618  us/op
CodecBenchmarks.readChineseSpan_proto3           avgt   15   1.293 ± 0.022  us/op
CodecBenchmarks.readChineseSpan_proto3_protobuf  avgt   15   0.075 ± 0.002  us/op
CodecBenchmarks.readClientSpan_java              avgt   15   5.362 ± 0.297  us/op
CodecBenchmarks.readClientSpan_json              avgt   15   4.535 ± 0.154  us/op
CodecBenchmarks.readClientSpan_proto3            avgt   15   1.191 ± 0.075  us/op
CodecBenchmarks.readClientSpan_proto3_protobuf   avgt   15   0.070 ± 0.004  us/op
CodecBenchmarks.readTenClientSpans_json          avgt   15  34.860 ± 0.463  us/op
CodecBenchmarks.writeChineseSpan_json            avgt   15   1.053 ± 0.062  us/op
CodecBenchmarks.writeChineseSpan_proto3          avgt   15   0.849 ± 0.021  us/op
CodecBenchmarks.writeClientSpan_java             avgt   15   2.056 ± 0.133  us/op
CodecBenchmarks.writeClientSpan_json             avgt   15   0.999 ± 0.024  us/op
CodecBenchmarks.writeClientSpan_json_legacy      avgt   15   2.109 ± 0.113  us/op
CodecBenchmarks.writeClientSpan_proto3           avgt   15   0.670 ± 0.015  us/op
CodecBenchmarks.writeTenClientSpans_json         avgt   15   9.523 ± 0.190  us/op
CodecBenchmarks.writeTenClientSpans_json_legacy  avgt   15  20.119 ± 0.468  us/op
```

after
```
Benchmark                                        Mode  Cnt   Score   Error  Units
CodecBenchmarks.readChineseSpan_json             avgt   15   4.395 ± 0.186  us/op
CodecBenchmarks.readChineseSpan_proto3           avgt   15   1.488 ± 0.053  us/op
CodecBenchmarks.readChineseSpan_proto3_protobuf  avgt   15   0.083 ± 0.010  us/op
CodecBenchmarks.readClientSpan_java              avgt   15   5.674 ± 0.197  us/op
CodecBenchmarks.readClientSpan_json              avgt   15   4.614 ± 0.241  us/op
CodecBenchmarks.readClientSpan_proto3            avgt   15   1.406 ± 0.167  us/op
CodecBenchmarks.readClientSpan_proto3_protobuf   avgt   15   0.076 ± 0.006  us/op
CodecBenchmarks.readTenClientSpans_json          avgt   15  35.355 ± 2.559  us/op
CodecBenchmarks.writeChineseSpan_json            avgt   15   1.039 ± 0.007  us/op
CodecBenchmarks.writeChineseSpan_proto3          avgt   15   0.846 ± 0.045  us/op
CodecBenchmarks.writeClientSpan_java             avgt   15   2.187 ± 0.184  us/op
CodecBenchmarks.writeClientSpan_json             avgt   15   1.118 ± 0.125  us/op
CodecBenchmarks.writeClientSpan_json_legacy      avgt   15   2.040 ± 0.144  us/op
CodecBenchmarks.writeClientSpan_proto3           avgt   15   0.688 ± 0.034  us/op
CodecBenchmarks.writeTenClientSpans_json         avgt   15  10.691 ± 0.410  us/op
CodecBenchmarks.writeTenClientSpans_json_legacy  avgt   15  20.426 ± 3.441  us/op
```